hacker_typer_txt.txt added as rescource (#1)
I have added the hacker_typer_txt.txt file as a resource to the program. 
This means that the file can be read on any computer.
If it throws any error please create an issue.
T54
P.S. Apologies for my bad grammar and spelling. Do comment to correct it.